chore: prepare VIP-631 for proposal
